Chat bug fix
if the user is deleted don't get the their but use -1 ( nil coalescing ) so that the comparison that determines which messages should be on the left or right side fails. The comparison is always false (not the logged in user) and the message is displayed on the right side, no avatar is fetched and no username is fetched.
If the user is deleted the parameters attribute in the threads array is empty, so at opening the chat view a boolean is sent to determine wether the bottom (reply to the thread) will be displayed (false if the user is deleted - parameters array is empty)